requirements.txt
Flask==3.0.3
yfinance==0.2.63
python-dateutil==2.9.0.post0
pytest==8.3.2
